# doc/swagger.yaml
definitions:
...
  crypto.PublicKeyJWK:
    properties:
      alg:
        type: string
      crv:
        type: string
      e:
        type: string
      key_ops:
        type: string
      kid:
        type: string
      kty:
        type: string
      "n":
        type: string
      use:
        type: string
      x:
        type: string
      "y":
        type: string
    required:
    - kty
    type: object
  did.DIDDocument:
    properties:
      '@context': {}
      alsoKnownAs:
        type: string
      assertionMethod:
        items: {}
        type: array
      authentication:
        items: {}
        type: array
      capabilityDelegation:
        items: {}
        type: array
      capabilityInvocation:
        items: {}
        type: array
      controller:
        type: string
      id:
        description: |-
          As per https://www.w3.org/TR/did-core/#did-subject intermediate representations of DID Documents do not
          require an ID property. The provided test vectors demonstrate IRs. As such, the property is optional.
        type: string
      keyAgreement:
        items: {}
        type: array
      service:
        items:
          $ref: '#/definitions/did.Service'
        type: array
      verificationMethod:
        items:
          $ref: '#/definitions/did.VerificationMethod'
        type: array
    type: object
  did.DIDDocumentMetadata:
    properties:
      canonicalId:
        type: string
      created:
        type: string
      deactivated:
        type: boolean
      equivalentId:
        type: string
      nextUpdate:
        type: string
      nextVersionId:
        type: string
      updated:
        type: string
      versionId:
        type: string
    type: object
  did.DIDResolutionMetadata:
    properties:
      contentType:
        type: string
      error:
        $ref: '#/definitions/did.ResolutionError'
    type: object
  did.ResolutionError:
    properties:
      code:
        type: string
      invalidDid:
        type: boolean
      notFound:
        type: boolean
      representationNotSupported:
        type: boolean
    type: object
  did.Service:
    type: object
  did.VerificationMethod:
    properties:
      blockchainAccountId:
        description: for PKH DIDs - https://github.com/w3c-ccg/did-pkh/blob/90b28ad3c18d63822a8aab3c752302aa64fc9382/did-pkh-method-draft.md
        type: string
      controller:
        type: string
      id:
        type: string
      publicKeyBase58:
        type: string
      publicKeyJwk:
        $ref: '#/definitions/crypto.PublicKeyJWK'
        description: must conform to https://datatracker.ietf.org/doc/html/rfc7517
      publicKeyMultibase:
        description: https://datatracker.ietf.org/doc/html/draft-multiformats-multibase-03
        type: string
      type:
        type: string
    required:
    - controller
    - id
    - type
    type: object
  exchange.ClaimFormat:
    properties:
      jwt:
        $ref: '#/definitions/exchange.JWTType'
      jwt_vc:
        $ref: '#/definitions/exchange.JWTType'
      jwt_vp:
        $ref: '#/definitions/exchange.JWTType'
      ldp:
        $ref: '#/definitions/exchange.LDPType'
      ldp_vc:
        $ref: '#/definitions/exchange.LDPType'
      ldp_vp:
        $ref: '#/definitions/exchange.LDPType'
    type: object
...
  exchange.InputDescriptor:
    properties:
      constraints:
        $ref: '#/definitions/exchange.Constraints'
      format:
        $ref: '#/definitions/exchange.ClaimFormat'
      group:
        description: Must match a grouping strings listed in the `from` values of
          a submission requirement rule
        items:
          type: string
        type: array
      id:
        description: Must be unique within the Presentation Definition
        type: string
      name:
        type: string
      purpose:
        description: Purpose for which claim's data is being requested
        type: string
    required:
    - constraints
    - id
    type: object
  exchange.JWTType:
    properties:
      alg:
        items:
          type: string
        type: array
    required:
    - alg
    type: object
  exchange.LDPType:
    properties:
      proof_type:
        items:
          type: string
        type: array
    required:
    - proof_type
    type: object
  exchange.PresentationDefinition:
    properties:
      format:
        $ref: '#/definitions/exchange.ClaimFormat'
      frame:
        description: https://identity.foundation/presentation-exchange/#json-ld-framing-feature
      id:
        type: string
      input_descriptors:
        items:
          $ref: '#/definitions/exchange.InputDescriptor'
        type: array
      name:
        type: string
      purpose:
        type: string
      submission_requirements:
        items:
          $ref: '#/definitions/exchange.SubmissionRequirement'
        type: array
    required:
    - id
    - input_descriptors
    type: object
...
  exchange.SubmissionRequirement:
    properties:
      count:
        minimum: 1
        type: integer
      from:
        type: string
      from_nested:
        items:
          $ref: '#/definitions/exchange.SubmissionRequirement'
        type: array
      max:
        type: integer
      min:
        type: integer
      name:
        type: string
      purpose:
        type: string
      rule:
        type: string
    required:
    - rule
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.CreateDIDByMethodRequest:
    properties:
      didWebId:
        type: string
      keyType:
        type: string
    required:
    - keyType
    type: object
  github.com_tbd54566975_ssi-service_pkg_server_router.CreateDIDByMethodResponse:
    properties:
      did:
        $ref: '#/definitions/did.DIDDocument'
      keyType:
        type: string
      privateKeyBase58:
        type: string
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.CreateManifestRequest:
    properties:
      description:
        type: string
      format:
        $ref: '#/definitions/exchange.ClaimFormat'
      issuerDid:
        type: string
      issuerName:
        type: string
      name:
        type: string
      outputDescriptors:
        items:
          $ref: '#/definitions/manifest.OutputDescriptor'
        type: array
      presentationDefinition:
        $ref: '#/definitions/exchange.PresentationDefinition'
    required:
    - format
    - issuerDid
    - outputDescriptors
    type: object
  github.com_tbd54566975_ssi-service_pkg_server_router.CreateManifestResponse:
    properties:
      credential_manifest:
        $ref: '#/definitions/manifest.CredentialManifest'
      manifestJwt:
        type: string
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.GetDIDByMethodResponse:
    properties:
      did:
        $ref: '#/definitions/did.DIDDocument'
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.GetDIDsByMethodResponse:
    properties:
      dids:
        items:
          $ref: '#/definitions/did.DIDDocument'
        type: array
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.GetManifestResponse:
    properties:
      credential_manifest:
        $ref: '#/definitions/manifest.CredentialManifest'
      id:
        type: string
      manifestJwt:
        type: string
    type: object
  github.com_tbd54566975_ssi-service_pkg_server_router.GetManifestsResponse:
    properties:
      manifests:
        items:
          $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.GetManifestResponse'
        type: array
    type: object
...
  github.com_tbd54566975_ssi-service_pkg_server_router.ResolveDIDResponse:
    properties:
      didDocument:
        $ref: '#/definitions/did.DIDDocument'
      didDocumentMetadata:
        $ref: '#/definitions/did.DIDDocumentMetadata'
      didResolutionMetadata:
        $ref: '#/definitions/did.DIDResolutionMetadata'
    type: object
...
  issuing.ClaimTemplates:
    properties:
      data:
        additionalProperties: {}
        description: |-
          Values may be json path like strings, or any other JSON primitive. Each entry will be used to come up with a
          claim about the credentialSubject in the credential that will be issued.
        type: object
    type: object
  issuing.CredentialTemplate:
    properties:
      data:
        $ref: '#/definitions/issuing.CredentialTemplateData'
        description: Date that will be used to determine credential claims.
      expiry:
        $ref: '#/definitions/issuing.TimeLike'
        description: Parameter to determine the expiry of the credential.
      id:
        description: ID corresponding to an OutputDescriptor.ID from the manifest.
        type: string
      revocable:
        description: Whether the credentials created should be revocable.
        type: boolean
      schema:
        description: ID of the CredentialSchema to be used for the issued credential.
        type: string
    type: object
  issuing.CredentialTemplateData:
    properties:
      claims:
        $ref: '#/definitions/issuing.ClaimTemplates'
        description: The set of information that will be used to create claims.
      credentialInputDescriptor:
        description: |-
          Optional. When present, it's the ID of the input descriptor in the application. Corresponds to one of the
          PresentationDefinition.InputDescriptors[].ID in the credential manifest.
        type: string
    type: object
  issuing.IssuanceTemplate:
    properties:
      credentialManifest:
        description: ID of the credential manifest that this template corresponds
          to.
        type: string
      credentials:
        description: Info required to create a credential from a credential application.
        items:
          $ref: '#/definitions/issuing.CredentialTemplate'
        type: array
      id:
        description: ID of this template.
        type: string
      issuer:
        description: ID of the issuer that will be issuing the credentials.
        type: string
    required:
    - credentialManifest
    - issuer
    type: object
  issuing.TimeLike:
    properties:
      duration:
        description: For a fixed offset from when it was issued.
        type: integer
      time:
        description: For fixed time in the future.
        type: string
    type: object
...
  manifest.CredentialManifest:
    properties:
      description:
        type: string
      format:
        $ref: '#/definitions/exchange.ClaimFormat'
      id:
        type: string
      issuer:
        $ref: '#/definitions/manifest.Issuer'
      name:
        type: string
      output_descriptors:
        items:
          $ref: '#/definitions/manifest.OutputDescriptor'
        type: array
      presentation_definition:
        $ref: '#/definitions/exchange.PresentationDefinition'
      spec_version:
        type: string
    required:
    - id
    - issuer
    - output_descriptors
    - spec_version
    type: object
...
  manifest.Issuer:
    properties:
      id:
        type: string
      name:
        type: string
      styles:
        $ref: '#/definitions/rendering.EntityStyleDescriptor'
        description: |-
          an object or URI as defined by the DIF Entity Styles specification
          https://identity.foundation/wallet-rendering/#entity-styles
    required:
    - id
    type: object
  manifest.OutputDescriptor:
    properties:
      description:
        type: string
      display:
        $ref: '#/definitions/rendering.DataDisplay'
        description: 'both below: an object or URI as defined by the DIF Entity Styles
          specification'
      id:
        description: Must be unique within a manifest
        type: string
      name:
        type: string
      schema:
        type: string
      styles:
        $ref: '#/definitions/rendering.EntityStyleDescriptor'
    required:
    - id
    - schema
    type: object
...
  pkg_server_router.CreateIssuanceTemplateRequest:
    properties:
      credentialManifest:
        description: ID of the credential manifest that this template corresponds
          to.
        type: string
      credentials:
        description: Info required to create a credential from a credential application.
        items:
          $ref: '#/definitions/issuing.CredentialTemplate'
        type: array
      id:
        description: ID of this template.
        type: string
      issuer:
        description: ID of the issuer that will be issuing the credentials.
        type: string
    required:
    - credentialManifest
    - issuer
    type: object
...
  rendering.ColorResource:
    properties:
      color:
        description: 'a HEX string color value (e.g. #00000)'
        type: string
    type: object
  rendering.DataDisplay:
    properties:
      description:
        $ref: '#/definitions/rendering.DisplayMappingObject'
      properties:
        items:
          $ref: '#/definitions/rendering.LabeledDisplayMappingObject'
        type: array
      subtitle:
        $ref: '#/definitions/rendering.DisplayMappingObject'
      title:
        $ref: '#/definitions/rendering.DisplayMappingObject'
    type: object
  rendering.DisplayMappingObject:
    properties:
      fallback:
        type: string
      path:
        description: |-
          Ifa path is present it must be an array of JSON Path string expressions
          and the schema property must also be present.
        items:
          type: string
        type: array
      schema:
        $ref: '#/definitions/rendering.DisplayMappingSchema'
      text:
        description: If path is not present, the text value is required with no other
          properties
        type: string
    type: object
...
  rendering.EntityStyleDescriptor:
    properties:
      background:
        $ref: '#/definitions/rendering.ColorResource'
      hero:
        $ref: '#/definitions/rendering.ImageResource'
      text:
        $ref: '#/definitions/rendering.ColorResource'
      thumbnail:
        $ref: '#/definitions/rendering.ImageResource'
    type: object
  rendering.ImageResource:
    properties:
      alt:
        description: Describes the alternate text for a logo image
        type: string
      uri:
        description: Must be a valid URI string to an image resource
        type: string
    required:
    - uri
    type: object
...
host: localhost:8080
info:
  contact:
    email: tbd-developer@squareup.com
    name: TBD
    url: https://github.com/TBD54566975/ssi-service/issues
  description: https://github.com/TBD54566975/ssi-service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SSI Service API
  version: "0.1"
paths:
...
  /v1/dids/{method}:
    get:
      consumes:
      - application/json
      description: Get DIDs by method
      parameters:
      - description: Method
        in: path
        name: method
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.GetDIDsByMethodResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Get DIDs
      tags:
      - DecentralizedIdentityAPI
    put:
      consumes:
      - application/json
      description: create DID by method
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.CreateDIDByMethodRequest'
      - description: Method
        in: path
        name: method
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.CreateDIDByMethodResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create DID
      tags:
      - DecentralizedIdentityAPI
  /v1/dids/{method}/{id}:
    get:
      consumes:
      - application/json
      description: Get DID by method
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.CreateDIDByMethodRequest'
      - description: Method
        in: path
        name: method
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.GetDIDByMethodResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Get DID
      tags:
      - DecentralizedIdentityAPI
  /v1/dids/resolver/{id}:
    get:
      consumes:
      - application/json
      description: Resolve a DID that may not be stored in this service
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.ResolveDIDResponse'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Resolve a DID
      tags:
      - DecentralizedIdentityAPI
  /v1/issuancetemplates:
    put:
      consumes:
      - application/json
      description: Create issuance template
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/pkg_server_router.CreateIssuanceTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/issuing.IssuanceTemplate'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create issuance template
      tags:
      - IssuingAPI
  /v1/issuancetemplates/{id}:
    delete:
      consumes:
      - application/json
      description: Delete issuance template by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Delete issuance template
      tags:
      - IssuingAPI
    get:
      consumes:
      - application/json
      description: Get an issuance template by its id
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/issuing.IssuanceTemplate'
        "400":
          description: Bad request
          schema:
            type: string
      summary: Get issuance template
      tags:
      - IssuingAPI
...
  /v1/manifests:
    get:
      consumes:
      - application/json
      description: Checks for the presence of a query parameter and calls the associated
        filtered get method
      parameters:
      - description: string issuer
        in: query
        name: issuer
        type: string
      - description: string schema
        in: query
        name: schema
        type: string
      - description: string subject
        in: query
        name: subject
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.GetManifestsResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get manifests
      tags:
      - ManifestAPI
    put:
      consumes:
      - application/json
      description: Create manifest
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.CreateManifestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github.com_tbd54566975_ssi-service_pkg_server_router.CreateManifestResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create manifest
      tags:
      - ManifestAPI
...
swagger: "2.0"